requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Google Generative AI
google-generativeai==0.3.1

# Data handling and formatting
pyarrow==14.0.1
jsonschema==4.20.0